reverting moist_cv, styling in map_single
